+++ libs/rxjs/array/src/index.ts
@@ -19,6 +19,10 @@ export { findAll } from './lib/find-all';
 export { findLast } from './lib/find-last';
 export { findIndex } from './lib/find-index';
 export { flipArray } from './lib/flip-array';
+export { fromMap } from './lib/from-map';
+export { fromObjectEntries } from './lib/from-object-entries';
+export { fromObjectKeys } from './lib/from-object-keys';
+export { fromSet } from './lib/from-set';
 export { indexOf } from './lib/index-of';
 export { intersects } from './lib/intersects';
 export { intersectsWith } from './lib/intersects-with';
@@ -29,5 +33,6 @@ export { shuffle } from './lib/shuffle';
 export { some } from './lib/some';
 export { sort } from './lib/sort';
 export { sortMap } from './lib/sort-map';
+export { toSet } from './lib/to-set';
 
 export { BinarySearchResult } from './types/binary-search';