+++ libs/rxjs/array/src/lib/from-object-keys.ts
@@ -1,0 +1,31 @@
+/**
+ * @packageDocumentation
+ * @module Array
+ */
+import { OperatorFunction } from 'rxjs';
+import { map } from 'rxjs/operators';
+
+/**
+ * Returns an Observable that emits an array from a source `Object` using Object.keys, the array contains
+ * the object keys as strings.
+ *
+ * @category Object
+ *
+ * @typeParam K The key type of the source Object
+ * @typeParam T The value type of the source Object
+ *
+ * @example Convert an Object into an array of keys
+ * ```ts
+ * const input = { 1: 'a', 2: 'b', 3: 'c' };
+ * of(input).pipe(fromObjectKeys()).subscribe();
+ * ```
+ * Output: `[ '1', '2', '3' ]`
+ *
+ * @returns Observable that emits a Array from a source Object keys
+ */
+export function fromObjectKeys<K extends string | number | symbol, T extends unknown>(): OperatorFunction<
+  Record<K, T>,
+  string[]
+> {
+  return (source) => source.pipe(map((value) => Object.keys(value)));
+}